+++ examples/richlist/richlist_form.cpp
@@ -51,8 +51,8 @@ void RichlistForm::InitWindow()
 
 bool RichlistForm::OnSelected(ui::EventArgs* args)
 {
-	int current = args->wParam;
-	int old = args->lParam;
+	int current = static_cast<int>(args->wParam);
+	int old = static_cast<int>(args->lParam);
 
 
 	auto message = nbase::StringPrintf(L"您选择了索引为 %d 的子项，上一次选择子项索引为 %d", current, old);
